Update the buy and sell section

+++ src/engine/stop_loss_engine.py
@@ -8,7 +8,6 @@ def start(user_id, position_id):
   print('Successfully started stop_loss engine')
   while True:
     position = database.get_position(user_id, position_id)
-    print(position)
     if position:
       market_cap = position['market_capital']
       token = position['token']
@@ -17,15 +16,18 @@ def start(user_id, position_id):
       current_market_data = token_engine.get_market_data(chain, token)
       print(current_market_data['market_capital'] / market_cap * 100)
       if market_cap > current_market_data['market_capital']:
-        
         if (1 - (current_market_data['market_capital'] / market_cap)) * 100 >= stop_loss:
-          swap_engine.sell(user_id, position_id, 100, 90)
-          print(f"stop_loss sell: {position_id}")
-          token_snipers = database.get_token_snipers(user_id, chain)
-          for index in range(len(token_snipers)):
-            if token_snipers[index]['is_auto'] == True:
-              database.remove_token_sniper(user_id, token_snipers[index]['id'])
-              break
+          tx_id, amount = swap_engine.sell(user_id, position_id, 100, 90)
+          if tx_id == None and amount == None:
+            tx_id, amount = swap_engine.sell(user_id, position_id, 100, 1000)
+          else:
+            print(f"stop_loss sell: {position_id}")
+            database.remove_position(user_id, position_id)
+            token_snipers = database.get_token_snipers(user_id, chain)
+            for index in range(len(token_snipers)):
+              if token_snipers[index]['is_auto'] == True:
+                database.remove_token_sniper(user_id, token_snipers[index]['id'])
+                break
     else:
       break
     time.sleep(3)
